Remove support for setup.py --compiler argument
