+++ extension_helpers/__init__.py
@@ -1,5 +1,4 @@
-from ._distutils_helpers import get_compiler
 from ._openmp_helpers import add_openmp_flags_if_available
-from ._setup_helpers import get_extensions, pkg_config
+from ._setup_helpers import get_compiler, get_extensions, pkg_config
 from ._utils import import_file, write_if_different
 from .version import version as __version__
